requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
python-socketio==5.14.3
aiohttp==3.13.2
python-dotenv==1.2.1
web3==6.15.1
eth-account==0.11.0
redis==5.0.1
sentry-sdk[flask]==1.40.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
gunicorn==21.2.0
gevent==23.9.1
bleach==6.1.0
eth-utils==2.3.1
requests==2.31.0

# LiveKit SDK
livekit==0.17.4
livekit-api==0.7.0
substrate-interface==1.7.11
